Edited

+++ app/models/fcdata.rb
@@ -631,10 +631,10 @@ class Fcdata < ApplicationRecord
       # E_PORPHYTRIN_T	E_PORPHYTRIN_U 평균값
       avr = self.e_porphyrin_u
 
-      avr1 = (Fcavgdata.where(age: avg_grade_1_field_name).first.e_porphyrin_u.to_i + Fcavgdata.where(age: avg_grade_1_field_name).first.e_porphyrin_t.to_i) / 2
-      avr2 = (Fcavgdata.where(age: avg_grade_2_field_name).first.e_porphyrin_u.to_i + Fcavgdata.where(age: avg_grade_2_field_name).first.e_porphyrin_t.to_i) / 2
-      avr3 = (Fcavgdata.where(age: avg_grade_3_field_name).first.e_porphyrin_u.to_i + Fcavgdata.where(age: avg_grade_3_field_name).first.e_porphyrin_t.to_i) / 2
-      avr4 = (Fcavgdata.where(age: avg_grade_4_field_name).first.e_porphyrin_u.to_i + Fcavgdata.where(age: avg_grade_4_field_name).first.e_porphyrin_t.to_i) / 2
+      avr1 = (Fcavgdata.where(age: avg_grade_1_field_name).first.e_porphyrin_u.to_f + Fcavgdata.where(age: avg_grade_1_field_name).first.e_porphyrin_t.to_f) / 2
+      avr2 = (Fcavgdata.where(age: avg_grade_2_field_name).first.e_porphyrin_u.to_f + Fcavgdata.where(age: avg_grade_2_field_name).first.e_porphyrin_t.to_f) / 2
+      avr3 = (Fcavgdata.where(age: avg_grade_3_field_name).first.e_porphyrin_u.to_f + Fcavgdata.where(age: avg_grade_3_field_name).first.e_porphyrin_t.to_f) / 2
+      avr4 = (Fcavgdata.where(age: avg_grade_4_field_name).first.e_porphyrin_u.to_f + Fcavgdata.where(age: avg_grade_4_field_name).first.e_porphyrin_t.to_f) / 2
 
       return convert_avg_to_five(avr: avr, avr1: avr1, avr2: avr2, avr3: avr3, avr4: avr4)
     end
@@ -655,6 +655,7 @@ class Fcdata < ApplicationRecord
   end
 
   def convert_avg_to_five(avr: avr, avr1: avr1, avr2: avr2, avr3: avr3, avr4: avr4)
+    #avr2 = 0.3 / avr3 = 0.7
     if avr < avr1
       return 4
     end
